requirements.txt
alabaster==0.7.10
appdirs==1.4.1
asn1crypto==0.22.0
astroid==1.5.3
attrs==15.2.0
Automat==0.0.0
Babel==2.5.0
backports.weakref==1.0rc1
bleach==1.5.0
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
decorator==4.1.2
docutils==0.14
entrypoints==0.2.3
Flask==0.12
Flask-HTTPAuth==3.2.3
gunicorn==19.6.0
html5lib==0.9999999
hyperlink==17.1.1
idna==2.6
imagesize==0.7.1
incremental==16.10.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==0.23
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.4.0
numpy==1.13.1
numpydoc==0.7.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.2
parsel==1.2.0
path.py==10.3.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.2.0
psutil==5.2.2
psycopg2==2.7.3.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pycparser==2.18
PyDispatcher==2.0.5
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.2
pyOpenSSL==17.0.0
pyparsing==2.1.10
python-dateutil==2.6.1
pytz==2017.2
pywin32==220
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
queuelib==1.4.2
requests==2.14.2
rope-py3k==0.9.4.post1
Scrapy==1.3.3
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
tensorflow-gpu==1.1.0
testpath==0.3
tornado==4.5.2
traitlets==4.3.2
Twisted==17.5.0
w3lib==1.17.0
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.11
zope.interface==4.4.2
